requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain-ollama
langchain_groq
fastapi
uvicorn
sse_starlette
langserve